requirement.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.24.86
botocore==1.27.86
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
decorator==5.1.1
Django==4.0.6
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-environ-2==2.3.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-moneyfield==0.2.1
django-mptt==0.14.0
django-recaptcha==3.0.0
django-richtextfield==1.6.1
django-storages==1.13.1
djangorestframework==3.14.0
environ==1.0
executing==1.2.0
gunicorn==19.6.0
idna==3.4
install==1.3.5
ipython==8.6.0
iyzipay==1.0.32
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.33
pure-eval==0.2.2
pycryptodome==3.16.0
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
sentry-sdk==1.9.10
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.1
traitlets==5.5.0
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.2.5
